+++ composables/useTeams.ts
@@ -1,18 +1,4 @@
-export interface iconInterface {
-  src: string
-  alt: string
-}
-export interface playerInterface {
-  playerName: string
-  position: string
-  belongs: string
-}
-export interface teamInterface {
-  teamId: number
-  teamName: string
-  icon: iconInterface
-  players: playerInterface[]
-}
+import { teamInterface } from './interfaces/teamInterfaces'
 
 export const useTeams = () => {
   const teamsState = useState<teamInterface>('teamsState', () => [])
